internal: Use yarn pnp
