+++ .yarn/sdks/prettier/package.json
@@ -1,0 +1,7 @@
+{
+  "name": "prettier",
+  "version": "3.4.2-sdk",
+  "main": "./index.cjs",
+  "type": "commonjs",
+  "bin": "./bin/prettier.cjs"
+}
