+++ .yarn/sdks/typescript/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "typescript",
+  "version": "5.7.2-sdk",
+  "main": "./lib/typescript.js",
+  "type": "commonjs",
+  "bin": {
+    "tsc": "./bin/tsc",
+    "tsserver": "./bin/tsserver"
+  }
+}
