+++ .yarnrc.yml
@@ -2,7 +2,8 @@ changesetIgnorePatterns:
   - "**/__tests__/**"
   - "**/*.md"
 
-nodeLinker: node-modules
+nodeLinker: pnp
+pnpEnableEsmLoader: true
 
 yarnPath: .yarn/releases/yarn-4.5.3.cjs
