+++ __tests__/new.ts
@@ -1,6 +1,3 @@
-import { Temporal } from '@js-temporal/polyfill';
-import React, { createContext, useContext } from 'react';
-
 import {
   schema,
   Endpoint,
@@ -16,6 +13,8 @@ import {
   Resource,
   ResourceOptions,
 } from '@data-client/rest';
+import { Temporal } from '@js-temporal/polyfill';
+import React, { createContext, useContext } from 'react';
 
 /** Represents data with primary key being from 'id' field. */
 export class IDEntity extends Entity {
