+++ examples/coin-app/package.json
@@ -36,7 +36,7 @@
     "@types/react": "*",
     "@types/react-dom": "*",
     "react-refresh": "*",
-    "typescript": "5.7",
+    "typescript": "5.7.2",
     "webpack": "*",
     "webpack-cli": "*"
   },
