+++ packages/core/package.json
@@ -126,6 +126,9 @@
   },
   "devDependencies": {
     "@anansi/browserslist-config": "^1.4.2",
-    "@types/node": "^22.0.0"
+    "@data-client/endpoint": "workspace:*",
+    "@types/jest": "^29.5.14",
+    "@types/node": "^22.0.0",
+    "rollup-plugins": "workspace:*"
   }
 }
