+++ packages/endpoint/package.json
@@ -135,7 +135,13 @@
   },
   "devDependencies": {
     "@anansi/browserslist-config": "^1.4.2",
+    "@data-client/core": "workspace:*",
+    "@data-client/normalizr": "workspace:*",
     "@js-temporal/polyfill": "^0.4.4",
-    "@types/node": "^22.0.0"
+    "@types/jest": "^29.5.14",
+    "@types/node": "^22.0.0",
+    "immutable": "4.2.2",
+    "nock": "^13.5.6",
+    "rollup-plugins": "workspace:*"
   }
 }
