+++ packages/endpoint/rollup.config.mjs
@@ -1,13 +1,6 @@
-import babel from 'rollup-plugin-babel';
-import commonjs from 'rollup-plugin-commonjs';
-import dts from 'rollup-plugin-dts';
-import filesize from 'rollup-plugin-filesize';
-import json from 'rollup-plugin-json';
-import resolve from 'rollup-plugin-node-resolve';
-import replace from 'rollup-plugin-replace';
-import { terser } from 'rollup-plugin-terser';
+import { babel, dts, commonjs, filesize, json, resolve, replace, terser } from 'rollup-plugins';
 
-import pkg from './package.json';
+import pkg from './package.json' with { type: 'json' };
 
 const dependencies = Object.keys(pkg.dependencies).filter(
   dep => !['@babel/runtime'].includes(dep),
