+++ packages/graphql/package.json
@@ -109,6 +109,9 @@
   },
   "devDependencies": {
     "@anansi/browserslist-config": "^1.4.2",
-    "@types/node": "^22.0.0"
+    "@types/jest": "^29.5.14",
+    "@types/node": "^22.0.0",
+    "nock": "^13.5.6",
+    "rollup-plugins": "workspace:*"
   }
 }
