+++ packages/graphql/rollup.config.mjs
@@ -1,42 +1,24 @@
-import babel from 'rollup-plugin-babel';
-import commonjs from 'rollup-plugin-commonjs';
-import filesize from 'rollup-plugin-filesize';
-import json from 'rollup-plugin-json';
-import resolve from 'rollup-plugin-node-resolve';
-import replace from 'rollup-plugin-replace';
-import { terser } from 'rollup-plugin-terser';
+import { babel, commonjs, filesize, json, resolve, replace, terser, typeConfig, typeConfigNext } from 'rollup-plugins';
 
-import pkg from './package.json';
+import pkg from './package.json' with { type: 'json' };
 
-const dependencies = Object.keys(pkg.dependencies)
-  .concat(Object.keys(pkg.peerDependencies))
-  .filter(dep => !['@babel/runtime'].includes(dep));
-const peers = Object.keys(pkg.peerDependencies);
+const dependencies = Object.keys(pkg.dependencies).filter(
+  dep => !['@babel/runtime'].includes(dep),
+);
 
 const extensions = ['.js', '.ts', '.tsx', '.mjs', '.json', '.node'];
 const nativeExtensions = ['.native.ts', ...extensions];
 process.env.NODE_ENV = 'production';
-process.env.BROWSERSLIST_ENV = 'legacy';
 
 function isExternal(id) {
   return dependencies.some(dep => dep === id || id.startsWith(dep));
 }
-
-export default [
+const configs = [];
+if (process.env.BROWSERSLIST_ENV !== 'node12') {
   // browser-friendly UMD build
-  {
+  configs.push({
     input: 'lib/index.js',
-    external: id => peers.some(dep => dep === id || id.startsWith(dep)),
-    output: [
-      {
-        file: pkg.unpkg,
-        format: 'umd',
-        name: 'Img',
-        globals: {
-          '@data-client/react': 'RDC',
-        },
-      },
-    ],
+    output: [{ file: pkg.unpkg, format: 'umd', name: 'GraphQL' }],
     plugins: [
       babel({
         exclude: ['node_modules/**', '/**__tests__/**'],
@@ -51,9 +33,11 @@ export default [
       terser({}),
       filesize({ showBrotliSize: true }),
     ],
-  },
+  });
+  configs.push(typeConfig);
+} else {
   // node-friendly commonjs build
-  {
+  configs.push({
     input: 'lib/index.js',
     external: isExternal,
     output: [{ file: pkg.main, format: 'cjs' }],
@@ -68,5 +52,6 @@ export default [
       resolve({ extensions: nativeExtensions }),
       commonjs({ extensions: nativeExtensions }),
     ],
-  },
-];
+  });
+}
+export default configs;
