+++ packages/img/package.json
@@ -85,8 +85,11 @@
   "devDependencies": {
     "@anansi/browserslist-config": "^1.4.2",
     "@data-client/react": "workspace:*",
+    "@testing-library/react": "^16.1.0",
+    "@types/jest": "^29.5.14",
     "@types/node": "^22.0.0",
     "@types/react": "19.0.2",
-    "react": "19.0.0"
+    "react": "^19.0.0",
+    "rollup-plugins": "workspace:*"
   }
 }
