+++ packages/img/rollup.config.mjs
@@ -1,43 +1,33 @@
-import babel from 'rollup-plugin-babel';
-import banner from 'rollup-plugin-banner2';
-import commonjs from 'rollup-plugin-commonjs';
-import filesize from 'rollup-plugin-filesize';
-import json from 'rollup-plugin-json';
-import resolve from 'rollup-plugin-node-resolve';
-import replace from 'rollup-plugin-replace';
-import { terser } from 'rollup-plugin-terser';
+import { babel, commonjs, filesize, json, resolve, replace, terser } from 'rollup-plugins';
 
-import pkg from './package.json';
+import pkg from './package.json' with { type: 'json' };
 
 const dependencies = Object.keys(pkg.dependencies)
   .concat(Object.keys(pkg.peerDependencies))
-  .concat(['@data-client/core', 'data-client'])
   .filter(dep => !['@babel/runtime'].includes(dep));
 const peers = Object.keys(pkg.peerDependencies);
 
 const extensions = ['.js', '.ts', '.tsx', '.mjs', '.json', '.node'];
 const nativeExtensions = ['.native.ts', ...extensions];
 process.env.NODE_ENV = 'production';
+process.env.BROWSERSLIST_ENV = 'legacy';
 
 function isExternal(id) {
   return dependencies.some(dep => dep === id || id.startsWith(dep));
 }
-const configs = [];
-if (process.env.BROWSERSLIST_ENV !== 'node12') {
+
+export default [
   // browser-friendly UMD build
-  configs.push({
+  {
     input: 'lib/index.js',
     external: id => peers.some(dep => dep === id || id.startsWith(dep)),
     output: [
       {
         file: pkg.unpkg,
         format: 'umd',
-        name: 'RDC.SSR',
+        name: 'Img',
         globals: {
           '@data-client/react': 'RDC',
-          '@data-client/redux': 'RDC.Redux',
-          redux: 'Redux',
-          react: 'React',
         },
       },
     ],
@@ -55,10 +45,9 @@ if (process.env.BROWSERSLIST_ENV !== 'node12') {
       terser({}),
       filesize({ showBrotliSize: true }),
     ],
-  });
-} else {
+  },
   // node-friendly commonjs build
-  configs.push({
+  {
     input: 'lib/index.js',
     external: isExternal,
     output: [{ file: pkg.main, format: 'cjs' }],
@@ -72,9 +61,6 @@ if (process.env.BROWSERSLIST_ENV !== 'node12') {
       replace({ 'process.env.CJS': 'true' }),
       resolve({ extensions: nativeExtensions }),
       commonjs({ extensions: nativeExtensions }),
-      // for nextjs 13 compatibility in node https://nextjs.org/docs/app/building-your-application/rendering
-      banner(() => "'use client';\n"),
     ],
-  });
-}
-export default configs;
+  },
+];
