+++ packages/normalizr/package.json
@@ -120,7 +120,11 @@
   },
   "devDependencies": {
     "@anansi/browserslist-config": "^1.4.2",
+    "@data-client/endpoint": "workspace:*",
+    "@js-temporal/polyfill": "^0.4.4",
+    "@types/jest": "^29.5.14",
     "@types/node": "^22.0.0",
-    "immutable": "4.2.2"
+    "immutable": "4.2.2",
+    "rollup-plugins": "workspace:*"
   }
 }
