+++ packages/normalizr/rollup.config.mjs
@@ -1,9 +1,4 @@
-import babel from 'rollup-plugin-babel';
-import commonjs from 'rollup-plugin-commonjs';
-import dts from 'rollup-plugin-dts';
-import filesize from 'rollup-plugin-filesize';
-import resolve from 'rollup-plugin-node-resolve';
-import { terser } from 'rollup-plugin-terser';
+import { babel, commonjs, filesize, dts, resolve, terser } from 'rollup-plugins';
 
 const name = 'normalizr';
