+++ packages/react/rollup.config.mjs
@@ -1,14 +1,6 @@
-import babel from 'rollup-plugin-babel';
-import banner from 'rollup-plugin-banner2';
-import commonjs from 'rollup-plugin-commonjs';
-import filesize from 'rollup-plugin-filesize';
-import json from 'rollup-plugin-json';
-import resolve from 'rollup-plugin-node-resolve';
-import replace from 'rollup-plugin-replace';
-import { terser } from 'rollup-plugin-terser';
+import { babel, banner, commonjs, filesize, json, resolve, replace, terser, typeConfig, typeConfigNext } from 'rollup-plugins';
 
-import pkg from './package.json';
-import { typeConfig, typeConfigNext } from '../../scripts/rollup-utils';
+import pkg from './package.json' with { type: 'json' };
 
 const dependencies = Object.keys(pkg.dependencies)
   .concat(Object.keys(pkg.peerDependencies))
