+++ packages/rest/package.json
@@ -136,6 +136,11 @@
   },
   "devDependencies": {
     "@anansi/browserslist-config": "^1.4.2",
-    "@types/node": "^22.0.0"
+    "@data-client/react": "workspace:*",
+    "@data-client/test": "workspace:*",
+    "@types/jest": "^29.5.14",
+    "@types/node": "^22.0.0",
+    "nock": "^13.5.6",
+    "rollup-plugins": "workspace:*"
   }
 }
