+++ packages/ssr/package.json
@@ -141,6 +141,7 @@
     "next": "^15.0.0",
     "react": "19.0.0",
     "react-dom": "19.0.0",
-    "redux": "^5.0.0"
+    "redux": "^5.0.0",
+    "rollup-plugins": "workspace:*"
   }
 }
