+++ packages/use-enhanced-reducer/package.json
@@ -87,7 +87,10 @@
   },
   "devDependencies": {
     "@anansi/browserslist-config": "^1.4.2",
+    "@types/jest": "^29.5.14",
     "@types/node": "^22.0.0",
-    "@types/react": "19.0.2"
+    "@types/react": "19.0.2",
+    "react": "^19.0.0",
+    "rollup-plugins": "workspace:*"
   }
 }
