+++ scripts/rollup-plugins/index.js
@@ -1,0 +1,10 @@
+export {default as babel} from 'rollup-plugin-babel';
+export {default as banner} from 'rollup-plugin-banner2';
+export {default as commonjs} from 'rollup-plugin-commonjs';
+export {default as resolve} from 'rollup-plugin-node-resolve';
+export {default as filesize} from 'rollup-plugin-filesize';
+export {default as json} from 'rollup-plugin-json';
+export {default as replace} from 'rollup-plugin-replace';
+export { terser } from 'rollup-plugin-terser';
+export {default as dts} from 'rollup-plugin-dts';
+export * from './rollup-utils.js'
