+++ website/package.json
@@ -56,7 +56,7 @@
     "react-dom": "^19.0.0",
     "react-json-tree": "0.19.0",
     "react-live": "^4.0.0",
-    "typescript": "5.7",
+    "typescript": "5.7.2",
     "uuid": "^11.0.0"
   },
   "browserslist": [
